Rename scripts and commands to be main asset agnostic
