+++ scripts/deployStablecoinPool.ts
