+++ scripts/test.ts
@@ -6,14 +6,14 @@ import {main as upgradeRebalancer} from "./upgradeRebalancer";
 import {main as upgradeRepayer} from "./upgradeRepayer";
 import {main as redeployStash} from "./redeployStash";
 import {main as deployCensoredMulticall} from "./deployCensoredMulticall";
-import {main as deployUSDCPool} from "./deployUSDCPool";
-import {main as deployUSDCStablecoinPool} from "./deployUSDCStablecoinPool";
+import {main as deployPool} from "./deployPool";
+import {main as deployStablecoinPool} from "./deployStablecoinPool";
 import {main as deployRepayer} from "./deployRepayer";
 import {main as upgradeLiquidityHub} from "./upgradeLiquidityHub";
-import {main as deployUSDCPoolAave} from "./deployUSDCPoolAave";
-import {main as deployUSDCPoolAaveLongTerm} from "./deployUSDCPoolAaveLongTerm";
+import {main as deployPoolAave} from "./deployPoolAave";
+import {main as deployPoolAaveLongTerm} from "./deployPoolAaveLongTerm";
 import {main as deployStandaloneRepayer} from "./deployStandaloneRepayer";
-import {main as deployUSDCPublicPool} from "./deployUSDCPublicPool";
+import {main as deployPublicPool} from "./deployPublicPool";
 import {main as deployERC4626Adapter} from "./deployERC4626Adapter";
 import {main as deployTokenProcessor} from "./deployTokenProcessor";
 import {main as upgradeTokenProcessor} from "./upgradeTokenProcessor";
@@ -33,18 +33,18 @@ async function main() {
   await upgradeRepayer();
   console.log("Test redeployStash.");
   await redeployStash();
-  console.log("Test deployUSDCPool.");
-  await deployUSDCPool();
-  console.log("Test deployUSDCStablecoinPool.");
-  await deployUSDCStablecoinPool();
+  console.log("Test deployPool.");
+  await deployPool();
+  console.log("Test deployStablecoinPool.");
+  await deployStablecoinPool();
   console.log("Test upgradeLiquidityHub.");
   await upgradeLiquidityHub();
-  console.log("Test deployUSDCPoolAave.");
-  await deployUSDCPoolAave();
-  console.log("Test deployUSDCPoolAaveLongTerm.");
-  await deployUSDCPoolAaveLongTerm();
-  console.log("Test deployUSDCPublicPool.");
-  await deployUSDCPublicPool();
+  console.log("Test deployPoolAave.");
+  await deployPoolAave();
+  console.log("Test deployPoolAaveLongTerm.");
+  await deployPoolAaveLongTerm();
+  console.log("Test deployPublicPool.");
+  await deployPublicPool();
   console.log("Test deployERC4626Adapter.");
   await deployERC4626Adapter();
   console.log("Test deployRepayer.");
